+++ src/Command/PrepareImages.fs
@@ -1,9 +1,11 @@
 namespace MF.ImageManager.Command
 
 open System.IO
-open ErrorHandling
-open MF.ImageManager
 open MF.ConsoleApplication
+open MF.ErrorHandling
+open MF.ImageManager
+open MF.ImageManager.Prepare
+open MF.Utils
 
 [<RequireQualifiedAccess>]
 module PrepareCommand =
@@ -17,87 +19,111 @@ module PrepareCommand =
         Option.optional "exclude-list" (Some "x") "Text file contains a list of files you want to exclude from searching (<c:yellow>one file at line</c>)." None
         Option.noValue "force" (Some "f") "If set, target directory will NOT be excluded, and images may be overwritten."
         Option.noValue "dry-run" None "If set, target directory will NOT be touched in anyway and images will only be sent to stdout."
+        Option.optional "prefix" None "Prefix all images with this static prefix, to allow the same image name from other sources." None
         Option.noValue "year" None "If set, target directory will have a sub-directory with year of the image created date."
         Option.noValue "month" None "If set, target directory will have a sub-directory with month of the image created date."
         Option.optional "fallback" None "If set, it will be used as a sub-directory in target directory for all files, which don't have other specific sub-directory." None
         Option.optional "config" (Some "c") "If set, config file will be used (other options set directly, will override a config values)." None
     ]
 
-    let execute = ExecuteResult <| fun (input, output) -> result {
-        let source = input |> Input.Option.asList "source"
-        let target = input |> Input.Argument.value "target"
-        let exclude =
-            match input with
-            | Input.Option.Has "exclude" _ ->
-                input |> Input.Option.asList "exclude"
-            | _ -> []
-        let excludeList =
-            match input with
-            | Input.Option.Has "exclude-list" _ ->
-                input |> Input.Option.asString "exclude-list"
-            | _ -> None
+    let execute ((input, output): IO) =
+        asyncResult {
+            let source = input |> Input.getOptionValueAsList "source"
+            let target = input |> Input.getArgumentValue "target"
+            let exclude =
+                match input with
+                | Input.HasOption "exclude" _ ->
+                    input |> Input.getOptionValueAsList "exclude"
+                | _ -> []
+            let excludeList =
+                match input with
+                | Input.HasOption "exclude-list" _ ->
+                    input |> Input.getOptionValueAsString "exclude-list"
+                | _ -> None
+            let prefix =
+                match input with
+                | Input.HasOption "prefix" (OptionValue.ValueOptional (Some value)) -> Some (Prefix value)
+                | _ -> None
 
-        let config =
             let targetDirMode =
                 match input with
-                | Input.Option.IsSet "force" _ -> Override
-                | Input.Option.IsSet "dry-run" _ -> DryRun
+                | Input.IsSetOption "force" _ -> Override
+                | Input.IsSetOption "dry-run" _ -> DryRun
                 | _ -> Exclude
 
-            match input with
-            | Input.Option.Has "config" (OptionValue.ValueOptional (Some config)) ->
+            let! parsedConfig =
+                match input with
+                | Input.HasOption "config" (OptionValue.ValueOptional (Some config)) ->
+                    config
+                    |> File.ReadAllText
+                    |> Config.parse targetDirMode
+                    |> Result.map Some
+                | _ -> Ok None
+                |> AsyncResult.ofResult
+                |> AsyncResult.mapError List.singleton
+
+            let config =
+                parsedConfig
+                |> Config.combine ({
+                    Source = source
+                    Prefix = prefix
+                    Target = target
+                    TargetDirMode = targetDirMode
+                    TargetSubdir =
+                        match input, input with
+                        | Input.IsSetOption "year" _, Input.IsSetOption "month" _ -> ByYearAndMonth
+                        | Input.IsSetOption "year" _, _ -> ByYear
+                        | Input.IsSetOption "month" _, _ -> ByMonth
+                        | _ -> Flat
+                    TargetSubdirFallback =
+                        match input with
+                        | Input.HasOption "fallback" fallback -> fallback |> OptionValue.stringValue
+                        | _ -> None
+                    Exclude =
+                        match exclude with
+                        | [] -> None
+                        | excludeDirs -> Some excludeDirs
+                    ExcludeList =
+                        match excludeList with
+                        | None -> None
+                        | Some excludeList ->
+                            if excludeList |> File.Exists |> not then
+                                failwithf "File %A you want to exclude from, does not exists." excludeList
+                            Some excludeList
+                    Ffmpeg = FFMpeg.Empty
+                })
+
+            if output.IsVerbose() then
+                output.Table ["Source"; "Prefix"; "Target"; "Exclude"; "Exclude list from"; "FFMpeg"]
+                    [
+                        [
+                            config.Source
+                            [config.Prefix |> sprintf "%A"]
+                            [config.Target]
+                            [config.Exclude |> sprintf "%A"]
+                            config.ExcludeList |> Option.toList
+                            [config.Ffmpeg |> sprintf "%A"]
+                        ]
+                        |> List.map (sprintf "%A")
+                    ]
+
+            output.Section <| sprintf "Prepare images to %s" config.Target
+            output.Message <| sprintf "From:\n - %s" (config.Source |> String.concat "\n - ")
+
+            return!
                 config
-                |> File.ReadAllText
-                |> PrepareForSorting.parse targetDirMode
-                |> Some
-            | _ -> None
-            |> PrepareForSorting.combine {
-                Source = source
-                Target = target
-                TargetDirMode = targetDirMode
-                TargetSubdir =
-                    match input, input with
-                    | Input.Option.IsSet "year" _, Input.Option.IsSet "month" _ -> ByYearAndMonth
-                    | Input.Option.IsSet "year" _, _ -> ByYear
-                    | Input.Option.IsSet "month" _, _ -> ByMonth
-                    | _ -> Flat
-                TargetSubdirFallback =
-                    match input with
-                    | Input.Option.Has "fallback" fallback -> fallback |> OptionValue.stringValue
-                    | _ -> None
-                Exclude =
-                    match exclude with
-                    | [] -> None
-                    | excludeDirs -> Some excludeDirs
-                ExcludeList =
-                    match excludeList with
-                    | None -> None
-                    | Some excludeList ->
-                        if excludeList |> File.Exists |> not then
-                            failwithf "File %A you want to exclude from, does not exists." excludeList
-                        Some excludeList
-                FfmpegPath = None   // todo - for now, it is only possible to set it by config file
-            }
-
-        if output.IsVerbose() then
-            output.Table ["Source"; "Target"; "Exclude"; "Exclude list from"]
-                [[source; [target]; exclude; excludeList |> Option.toList] |> List.map (sprintf "%A")]
-
-        output.Section <| sprintf "Prepare images to %s" target
-        output.Message <| sprintf "From:\n - %s" (source |> String.concat "\n - ")
-
-        let! message =
-            config
-            |> Prepare.prepareForSorting output
-
-        output.Success message
-        (* |> function
+                |> Prepare.prepareForSorting output
+        }
+        |> Async.RunSynchronously
+        |> function
             | Ok message ->
                 output.Success message
                 ExitCode.Success
-            | Error message ->
-                output.Error message
-                ExitCode.Error *)
+            | Error errors ->
+                errors
+                |> List.iter (function
+                    | PrepareError.Exception e -> output.Error e.Message
+                    | PrepareError.ErrorMessage message -> output.Error message
+                )
 
-        return ExitCode.Success
-    }
+                ExitCode.Error